+++ packages/client-connectivity/src/sdk/bankaccounts.ts
@@ -213,9 +213,8 @@ export class BankAccounts extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }