+++ packages/client-connectivity/src/sdk/connectors.ts
@@ -610,9 +610,8 @@ export class Connectors extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }