+++ packages/client-connectivity/src/sdk/paymentsv1.ts
@@ -215,9 +215,8 @@ export class PaymentsV1 extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }