+++ packages/client-connectivity/src/sdk/transferinitiations.ts
@@ -139,9 +139,8 @@ export class TransferInitiations extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }